ml-service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
numpy
psycopg2-binary
python-dotenv
sqlalchemy
seaborn


